tienda/requirements.txt
asgiref==3.7.2
Django==4.2.7
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3